Move software_auto client to lib/
